Apply a workaround for odd FreeBSD recvmsg behavior
The recvmsg syscall in FreeBSD seems to be following the semantics
of read, where attempting to read zero bytes is a no-op. When there is
actually a zero length message on the socket, recvmsg sets the MSG_TRUNC
flag in msghdr.msg_flags, so this can be used to differentiate between
the no-message, and the zero-length-message scenarios.